feat(form): edit articles' feature added

+++ src/index.js
@@ -2,8 +2,9 @@ import "./assets/scripts/topbar.js";
 import "./assets/styles/styles.scss";
 import "./index.scss";
 
-
 const articleContainerElement = document.querySelector(".articles-container");
+
+
 
 const createArticles = (articles) => {
     const articlesDOM = articles.map((article) => {
@@ -20,30 +21,41 @@ const createArticles = (articles) => {
 <p class="article-content">${article.content}</p>
 <div class="article-actions">
   <button class="btn btn-danger" data-id=${article._id}>Supprimer</button>
+  <button class="btn btn-primary" data-id=${article._id}>Modifier</button>
 </div>`
         return articleDOM;
     });
-    articleContainerElement.innerHTML = '';
-    articleContainerElement.append(...articlesDOM);
-    const deleteButtons = articleContainerElement.querySelectorAll('.btn-danger');
-    deleteButtons.forEach(button => {
-        button.addEventListener("click", async event => {
-            try {
-                const target = event.target;
+    if (articleContainerElement) {
+        articleContainerElement.innerHTML = '';
+        articleContainerElement.append(...articlesDOM);
+        const deleteButtons = articleContainerElement.querySelectorAll('.btn-danger');
+        const editButtons = articleContainerElement.querySelectorAll('.btn-primary');
+
+        editButtons.forEach(button => {
+            button.addEventListener("click", (e) => {
+                const target = e.target;
                 const articleID = target.dataset.id;
-                const response = await fetch(`https://restapi.fr/api/articles/${articleID}`, {
-                    method: "DELETE"
-                });
-                const body = await response.json();
-                // quand on supprime un article, on relance l'affichage de la liste des articles mis à jour :
-                fetchArticles();
-            } catch (error) {
-                console.log('e : ', error);
-            }
+                location.assign(`/form/form.html?id=${articleID}`); // rediriger vers le form, pour la modification, avec l'id dans l'url
+            })
+        });
 
+        deleteButtons.forEach(button => {
+            button.addEventListener("click", async event => {
+                try {
+                    const target = event.target;
+                    const articleID = target.dataset.id;
+                    const response = await fetch(`https://restapi.fr/api/articles/${articleID}`, {
+                        method: "DELETE"
+                    });
+                    const body = await response.json();
+                    // quand on supprime un article, on relance l'affichage de la liste des articles mis à jour :
+                    fetchArticles();
+                } catch (error) {
+                    console.log('e : ', error);
+                }
+            })
         })
-    })
-
+    }
 };
 
 // Récupérer la liste de nos articles : 
@@ -51,7 +63,6 @@ const fetchArticles = async () => {
     try {
         const response = await fetch("https://restapi.fr/api/articles");
         let articles = await response.json();
-        console.log(articles);
         // Restapi retourne un objet s'il n'y a qu'un seul article, nous devons donc le transformer en tableau :
         if (!Array.isArray(articles)) {
             articles = [articles];
